--- src/brain_output/mod.rs
@@ -1,5 +1,0 @@
-//! Brain output processing modules for motor control and actuator data in FEAGI.
-
-
-// Module for all brain output (primarily motor devices)
-pub mod relative_servo_output;
