+++ src/data_types/image/mod.rs
@@ -1,0 +1,13 @@
+/// Segmented vision processing for region-based image analysis.
+/// 
+/// Provides functionality for processing images into a central segment and peripheral
+/// vision segments.
+pub mod segmented_vision_frame;
+
+/// Visual feature descriptors and extraction helpers.
+pub mod descriptors;
+
+/// Core image frame data structure and processing functions.
+pub mod image_frame;
+
+
