+++ src/data_types/image/quick_image_diff.rs.disabled
