+++ src/data_types/mod.rs
@@ -1,0 +1,6 @@
+mod image;
+pub mod neuron_data;
+
+pub use image::image_frame::ImageFrame as ImageFrame;
+pub use image::segmented_vision_frame::SegmentedVisionFrame as SegmentedVisionFrame;
+pub use image::descriptors as image_descriptors;
