+++ src/data_types/neuron_data/mod.rs
@@ -1,0 +1,5 @@
+mod xyzp;
+
+pub use xyzp::NeuronXYZP as NeuronXYZP;
+pub use xyzp::NeuronXYZPArrays as NeuronXYZPArrays;
+pub use xyzp::CorticalMappedXYZPNeuronData as CorticalMappedXYZPNeuronData;
