+++ src/data_types/neuron_data/xyzp/mod.rs
@@ -1,0 +1,9 @@
+mod neuron_xyzp;
+mod neuron_xyzp_arrays;
+mod neuron_xyzp_mappings;
+
+pub use neuron_xyzp::NeuronXYZP as NeuronXYZP;
+
+pub use neuron_xyzp_arrays::NeuronXYZPArrays as NeuronXYZPArrays;
+
+pub use neuron_xyzp_mappings::CorticalMappedXYZPNeuronData as CorticalMappedXYZPNeuronData;
