--- src/neuron_data/mod.rs
@@ -1,4 +1,0 @@
-pub mod neuron_arrays;
-pub mod neuron_mappings;
-pub mod neurons;
-
